// src/views/insights/facebook/posts/interface.ts
/**
 **********************************************************************
 * This source code is shared for the sole purpose of demo. Do not copy!
 **********************************************************************
 **/

export interface IPostPerformance {
  id: string;
  message: string;
  createdTime: string;
  fullPicture: string;
  reach: number;
  engagement: number;
  reactions: number;
  clicks: number;
}